Fix compat with preloading
